+++ ChangeLog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Using custom glTF shaders. This speeds up occlusion and roughness/metallic texture loading since they don't have to be converted at runtime anymore. It's also the foundation for future material features.
+
 ## [0.7.1] - 2019-11-29
 ### Fixed
 - glTF binary with Draco compression (decoding error due to invalid buffer view access)